# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlightGo Backend API",
    "version": "1.0.0",
    "description": "FlightGo Backend API"
  },
  "servers": [
    {
      "url": "undefined",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SeatInfo": {
        "type": "object",
        "properties": {
          "flight_id": {
            "type": "integer",
            "description": "The flight ID"
          },
          "seat_number": {
            "type": "integer",
            "description": "The seat number"
          },
          "seat_status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE",
              "BOOKED"
            ],
            "description": "The status of the seat"
          },
          "version": {
            "type": "integer",
            "description": "Version for optimistic concurrency"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/seat": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new seat info",
        "tags": [
          "SeatInfo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "flight_id": {
                    "type": "integer",
                    "description": "The flight ID"
                  },
                  "seat_number": {
                    "type": "integer",
                    "description": "The seat number"
                  },
                  "seat_status": {
                    "type": "string",
                    "enum": [
                      "AVAILABLE",
                      "UNAVAILABLE",
                      "BOOKED"
                    ],
                    "description": "The seat status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seat info created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or seat number exceeds capacity"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "404": {
            "description": "Flight not found"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all seat infos",
        "tags": [
          "SeatInfo"
        ],
        "responses": {
          "200": {
            "description": "List of seat infos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          }
        }
      }
    },
    "/seat/{flight_id}/{seat_number}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a seat info",
        "tags": [
          "SeatInfo"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "flight_id",
            "required": true,
            "description": "The flight ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "seat_number",
            "required": true,
            "description": "The seat number",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Seat info found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "404": {
            "description": "Seat info not found"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a seat info",
        "tags": [
          "SeatInfo"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "flight_id",
            "required": true,
            "description": "The flight ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "seat_number",
            "required": true,
            "description": "The seat number",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seat_status": {
                    "type": "string",
                    "enum": [
                      "AVAILABLE",
                      "UNAVAILABLE",
                      "BOOKED"
                    ]
                  },
                  "version": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seat info updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or cannot update booked seat"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "404": {
            "description": "Seat info not found"
          },
          "409": {
            "description": "Version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "currentVersion": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a seat info",
        "tags": [
          "SeatInfo"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "flight_id",
            "required": true,
            "description": "The flight ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "seat_number",
            "required": true,
            "description": "The seat number",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Seat info deleted successfully"
          },
          "400": {
            "description": "Cannot delete seat info with associated tickets"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "404": {
            "description": "Seat info not found"
          }
        }
      }
    },
...
  },
...
}